requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-testdox==3.1.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
